Summary period controller
* Added perod summary controller.
* Removed expanded summary controller.
* Updated TCO OPTIONS method.
